--- a/conch/ops/vision/bev_pool.py
+++ b/conch/ops/vision/bev_pool.py
@@ -5,7 +5,7 @@
 
 import torch
 
-from conch.kernels.vision.bev_pool import bev_pool_launcher
+from conch.kernels.vision.bev_pool import bev_pool_backward_launcher, bev_pool_launcher
 
 
 def bev_pool(
@@ -51,3 +51,36 @@ def bev_pool(
     )
 
     return output
+
+
+def bev_pool_backward(
+    grad_output: torch.Tensor,
+    geom_feats: torch.Tensor,
+    interval_starts: torch.Tensor,
+    interval_lengths: torch.Tensor,
+) -> torch.Tensor:
+    """Backward pass for the BEV pooling operation.
+
+    Args:
+        grad_output: gradient of the output, FloatTensor[batch_size, grid_z_cells, grid_x_cells, grid_y_cells, c]
+        geom_feats: input coordinates, IntTensor[n, 4]
+        interval_starts: starting position for pooled point
+        interval_lengths: how many points in each pooled point
+
+    Returns:
+        Gradient with respect to image features.
+    """
+    num_points, _ = geom_feats.shape
+    _, _, _, _, num_channels = grad_output.shape
+
+    x_grad = torch.zeros((num_points, num_channels), dtype=grad_output.dtype, device=grad_output.device)
+
+    bev_pool_backward_launcher(
+        x_grad=x_grad,
+        grad_output=grad_output,
+        geom_feats=geom_feats,
+        interval_starts=interval_starts,
+        interval_lengths=interval_lengths,
+    )
+
+    return x_grad